pyproject.toml
[build-system]
requires = ["hatchling>=1.25,<2.0"]
build-backend = "hatchling.build"

[project]
name = "pyxle-langkit"
version = "0.2.0"
description = "Language tools for Pyxle: LSP server, linter, and editor integrations."
readme = "README.md"
authors = [
  { name = "Shivam Saini", email = "dev@pyxle.dev" }
]
license = {text = "MIT"}
requires-python = ">=3.10"
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Software Development :: Quality Assurance",
  "Topic :: Text Editors :: Integrated Development Environments (IDE)",
  "Typing :: Typed",
]
keywords = ["pyxle", "lsp", "linter", "vscode", "language-server"]
dependencies = [
  "pyxle-framework>=0.1.0",
  "pygls>=1.3,<2.0",
  "lsprotocol>=2023.0.0",
  "pyflakes>=3.2,<3.3",
  "jedi>=0.19,<0.20",
]

[project.urls]
Homepage = "https://pyxle.dev"
Repository = "https://github.com/pyxle-framework/pyxle-langkit"

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
  "pytest-cov>=4.1,<5.0",
  "coverage>=7.4,<8.0",
  "ruff>=0.6,<0.7",
]

[project.scripts]
pyxle-langkit = "pyxle_langkit.cli:app"
pyxle-langserver = "pyxle_langkit.server:main"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["pyxle_langkit"]

[tool.coverage.run]
branch = true
source = ["pyxle_langkit"]

[tool.coverage.report]
show_missing = true
skip_covered = true
fail_under = 80

[tool.pytest.ini_options]
addopts = "--strict-markers --strict-config --cov=pyxle_langkit --cov-report=term-missing"
testpaths = ["tests"]
pythonpath = ["."]
